+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next version
 
+- Add `paid_discount`, `paid_inretest`, `paid_mulct`, `paid_rebate` to Charge entity
+- Add `paid_difference` (vitual attribute) to Charge entity
+
 ## v0.16.0
 
 - Add `auto_send_billet` to Charge entity